Move toolchains to jvm test suite plugin
