+++ platforms/jvm/plugins-jvm-test-suite/src/main/java/org/gradle/api/plugins/jvm/internal/DefaultJvmTestSuite.java
@@ -24,20 +24,20 @@ import org.gradle.api.artifacts.ConfigurationContainer;
 import org.gradle.api.internal.artifacts.dsl.dependencies.DefaultDependencyAdder;
 import org.gradle.api.internal.tasks.JvmConstants;
 import org.gradle.api.internal.tasks.TaskDependencyFactory;
-import org.gradle.api.internal.testing.toolchains.FrameworkCachingJvmTestToolchain;
-import org.gradle.api.internal.testing.toolchains.JUnit4TestToolchain;
-import org.gradle.api.internal.testing.toolchains.JUnit4ToolchainParameters;
-import org.gradle.api.internal.testing.toolchains.JUnitJupiterTestToolchain;
-import org.gradle.api.internal.testing.toolchains.JUnitJupiterToolchainParameters;
-import org.gradle.api.internal.testing.toolchains.JvmTestToolchain;
-import org.gradle.api.internal.testing.toolchains.JvmTestToolchainParameters;
-import org.gradle.api.internal.testing.toolchains.KotlinTestTestToolchain;
-import org.gradle.api.internal.testing.toolchains.KotlinTestToolchainParameters;
-import org.gradle.api.internal.testing.toolchains.LegacyJUnit4TestToolchain;
-import org.gradle.api.internal.testing.toolchains.SpockTestToolchain;
-import org.gradle.api.internal.testing.toolchains.SpockToolchainParameters;
-import org.gradle.api.internal.testing.toolchains.TestNGTestToolchain;
-import org.gradle.api.internal.testing.toolchains.TestNGToolchainParameters;
+import org.gradle.api.testing.toolchains.internal.FrameworkCachingJvmTestToolchain;
+import org.gradle.api.testing.toolchains.internal.JUnit4TestToolchain;
+import org.gradle.api.testing.toolchains.internal.JUnit4ToolchainParameters;
+import org.gradle.api.testing.toolchains.internal.JUnitJupiterTestToolchain;
+import org.gradle.api.testing.toolchains.internal.JUnitJupiterToolchainParameters;
+import org.gradle.api.testing.toolchains.internal.JvmTestToolchain;
+import org.gradle.api.testing.toolchains.internal.JvmTestToolchainParameters;
+import org.gradle.api.testing.toolchains.internal.KotlinTestTestToolchain;
+import org.gradle.api.testing.toolchains.internal.KotlinTestToolchainParameters;
+import org.gradle.api.testing.toolchains.internal.LegacyJUnit4TestToolchain;
+import org.gradle.api.testing.toolchains.internal.SpockTestToolchain;
+import org.gradle.api.testing.toolchains.internal.SpockToolchainParameters;
+import org.gradle.api.testing.toolchains.internal.TestNGTestToolchain;
+import org.gradle.api.testing.toolchains.internal.TestNGToolchainParameters;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.plugins.JvmTestSuitePlugin;
 import org.gradle.api.plugins.jvm.JvmComponentDependencies;
