+++ platforms/jvm/plugins-jvm-test-suite/src/main/java/org/gradle/api/testing/toolchains/internal/JUnitPlatformToolchainParameters.java
@@ -14,15 +14,18 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.testing.toolchains;
+package org.gradle.api.testing.toolchains.internal;
 
 import org.gradle.api.provider.Property;
 
 /**
- * Parameters for configuring a KotlinTest test toolchain.
+ * Parameters for configuring a JUnit Platform test toolchain.
  *
  * @since 8.5
  */
-public interface KotlinTestToolchainParameters extends JUnitPlatformToolchainParameters {
-    Property<String> getKotlinTestVersion();
+public interface JUnitPlatformToolchainParameters extends JvmTestToolchainParameters {
+    /**
+     * The version of the JUnit Platform to use for executing tests.
+     */
+    Property<String> getPlatformVersion();
 }
